+++ s/transports/websocket/client.ts
@@ -25,7 +25,7 @@ export async function webSocketRemote<F extends Fns>(
 		getLocalEndpoint = () => null,
 	} = params
 
-	return await deadline("web socket remote", timeout, async() => {
+	return await deadline(timeout, "web socket remote", async() => {
 		const socket = new WebSocket(url)
 
 		const ready = new Promise<WebSocket>((resolve, reject) => {